requirements.txt
-e .
pytest
flake8
autopep8
python-benedict